Add test for search endpoint
These changes also do a lot of refactoring of the service in general to support proper test mocking. The biggest changes are: using pydantic's feature of reading .evn files to supply configuration, making more extensive use of dependencies (so they can be overriden by tests), and having the OAK implementation object be passed to the service methods.